Update overview.md
DELETED
If you’re in the UK, you can:

- [fill in your information online](/apply-renew-passport) then print and post your form, photos and documents to the address on the form (you can't use the Post Office Passport Check and Send service if you fill in the form online)
- get a paper form from a [Post Office that offers the Passport Check and Send service](http://www.postoffice.co.uk/passport-check-send)
- [contact the Passport Adviceline](/passport-advice-line) to get a form posted to you

REASON
This content sent users to apply before they could read the guide - applying for a child passport's quite complicated so we shouldn't give the impression you don't need to read the guide. 

This is now in a new chapter, 'How to apply'.